refactor(collector-core): split trigger.rs into trigger/ submodules
Pure mechanical module extraction with no logic, behavior, or public-API
changes. The monolithic trigger.rs (2793 lines) is split into:

- trigger/mod.rs (22 lines): module doc + re-exports preserving the full
  public API surface at crate::trigger::*
- trigger/types.rs (331 lines): config, condition, deduplication-key,
  metadata, process-data, capabilities, aggregate-statistics, and the
  TriggerError enum
- trigger/queue.rs (161 lines): PriorityTriggerQueue + QueueStatistics
- trigger/sql_evaluator.rs (395 lines): SqlTriggerEvaluator and its
  compiled-condition / evaluation-stats types
- trigger/manager.rs (1966 lines): the lock_or_err! macro, RateLimitState,
  the single TriggerManager impl, TriggerTimeout, plus co-located tests

manager.rs remains large because it is one type's single impl (~1095 lines
non-test) plus tests that access TriggerManager's private timeout_tracker
field and the private TriggerTimeout struct; those tests must stay
co-located, and sharding one impl would create arbitrary boundaries.

Test count unchanged (147 lib tests, 23 in trigger; doctests pass).
