+++ collector-core/src/trigger/manager.rs
@@ -1,9 +1,8 @@
-//! Trigger system for analysis collector coordination.
+//! Trigger system manager for analysis collector coordination.
 //!
-//! This module provides the infrastructure for coordinating between process
-//! monitoring collectors and analysis collectors (Binary Hasher, Memory Analyzer, etc.).
-//! It implements trigger condition evaluation, deduplication, rate limiting, and
-//! metadata tracking for forensic analysis.
+//! This submodule contains [`TriggerManager`], the central coordinator that
+//! evaluates trigger conditions, applies deduplication and rate limiting,
+//! emits trigger requests through the event bus, and tracks request timeouts.
 
 #![allow(clippy::significant_drop_tightening)]
 #![allow(clippy::as_conversions)]
@@ -29,151 +28,21 @@ macro_rules! lock_or_err {
     };
 }
 
-use crate::event::{AnalysisType, CollectionEvent, TriggerPriority, TriggerRequest};
+use crate::event::{CollectionEvent, TriggerRequest};
 use crate::event_bus::EventBus;
-use serde::{Deserialize, Serialize};
-use sqlparser::{dialect::GenericDialect, parser::Parser};
-use std::collections::{HashMap, VecDeque};
+use std::collections::HashMap;
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, SystemTime};
 use tokio::sync::RwLock;
 use tracing::{debug, info, warn};
 use uuid::Uuid;
 
-/// Configuration for the trigger system.
-///
-/// Controls trigger generation behavior, rate limiting, and deduplication
-/// to prevent analysis collector overload while ensuring critical threats
-/// are analyzed promptly.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerConfig {
-    /// Maximum number of triggers per collector per time window
-    pub max_triggers_per_collector: usize,
-
-    /// Time window for rate limiting (in seconds)
-    pub rate_limit_window_secs: u64,
-
-    /// Deduplication window for identical triggers (in seconds)
-    pub deduplication_window_secs: u64,
-
-    /// Maximum number of pending triggers to track
-    pub max_pending_triggers: usize,
-
-    /// Enable trigger metadata collection for debugging
-    pub enable_metadata_tracking: bool,
-
-    /// Default timeout for trigger requests (in seconds)
-    pub default_timeout_secs: u64,
-}
-
-impl Default for TriggerConfig {
-    fn default() -> Self {
-        Self {
-            max_triggers_per_collector: 100,
-            rate_limit_window_secs: 60,
-            deduplication_window_secs: 300, // 5 minutes
-            max_pending_triggers: 1000,
-            enable_metadata_tracking: true,
-            default_timeout_secs: 30,
-        }
-    }
-}
-
-/// Trigger condition for evaluating when to request analysis.
-///
-/// Defines the criteria that must be met to trigger analysis collector
-/// coordination. Conditions can be combined using logical operators.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerCondition {
-    /// Condition identifier
-    pub id: String,
-
-    /// Human-readable description
-    pub description: String,
-
-    /// Analysis type to trigger
-    pub analysis_type: AnalysisType,
-
-    /// Priority level for triggered analysis
-    pub priority: TriggerPriority,
-
-    /// Target collector name
-    pub target_collector: String,
-
-    /// Condition evaluation function (simplified for this implementation)
-    pub condition_type: ConditionType,
-}
-
-/// Types of trigger conditions that can be evaluated.
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-#[non_exhaustive]
-pub enum ConditionType {
-    /// Process name matches pattern
-    ProcessNamePattern(String),
-
-    /// Executable path matches pattern
-    ExecutablePathPattern(String),
-
-    /// Process has no executable file
-    MissingExecutable,
-
-    /// Process hash mismatch detected
-    HashMismatch,
-
-    /// Suspicious parent-child relationship
-    SuspiciousParentChild,
-
-    /// High resource usage anomaly
-    ResourceAnomaly {
-        cpu_threshold: f64,
-        memory_threshold: u64,
-    },
-
-    /// Custom SQL condition (not yet supported, always evaluates to false)
-    /// See issue #000 for tracking
-    Custom(String),
-}
-
-/// Deduplication key for trigger requests.
-///
-/// Used to identify identical or similar trigger requests to prevent
-/// redundant analysis of the same target within the deduplication window.
-#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
-pub struct DeduplicationKey {
-    /// Target collector name
-    pub collector: String,
-
-    /// Analysis type
-    pub analysis_type: AnalysisType,
-
-    /// Target process ID (if applicable)
-    pub target_pid: Option<u32>,
-
-    /// Target file path (if applicable)
-    pub target_path: Option<String>,
-}
-
-/// Trigger metadata for correlation and debugging.
-///
-/// Tracks trigger generation context, evaluation results, and timing
-/// information for forensic analysis and system debugging.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerMetadata {
-    /// Trigger generation timestamp
-    pub generated_at: SystemTime,
-
-    /// Source event that triggered the analysis
-    pub source_event_type: String,
-
-    /// Condition that matched
-    pub matched_condition: String,
-
-    /// Evaluation context
-    pub evaluation_context: HashMap<String, String>,
-
-    /// Correlation ID for tracking related events
-    pub correlation_id: String,
-}
+use super::queue::PriorityTriggerQueue;
+use super::sql_evaluator::SqlTriggerEvaluator;
+use super::types::{
+    ConditionType, DeduplicationKey, ProcessTriggerData, TriggerCapabilities, TriggerCondition,
+    TriggerConfig, TriggerEmissionStats, TriggerError, TriggerMetadata, TriggerStatistics,
+};
 
 /// Rate limiting state for a specific collector.
 #[derive(Debug, Clone)]
@@ -1280,658 +1149,6 @@ impl TriggerManager {
     }
 }
 
-/// Process data for trigger evaluation.
-///
-/// Contains the process information needed to evaluate trigger conditions
-/// and generate analysis requests.
-#[derive(Debug, Clone)]
-pub struct ProcessTriggerData {
-    /// Process ID
-    pub pid: u32,
-
-    /// Process name
-    pub name: String,
-
-    /// Executable path
-    pub executable_path: Option<String>,
-
-    /// Whether executable file exists
-    pub file_exists: bool,
-
-    /// CPU usage percentage
-    pub cpu_usage: Option<f64>,
-
-    /// Memory usage in bytes
-    pub memory_usage: Option<u64>,
-
-    /// Process hash (if available)
-    pub executable_hash: Option<String>,
-}
-
-/// Trigger capabilities advertised by collectors.
-///
-/// This structure defines the trigger conditions that a collector can evaluate
-/// and the analysis types it can perform. These capabilities are advertised
-/// during collector registration and used for trigger condition validation.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerCapabilities {
-    /// Collector identifier
-    pub collector_id: String,
-
-    /// Supported trigger condition types
-    pub supported_conditions: Vec<ConditionType>,
-
-    /// Supported analysis types
-    pub supported_analysis: Vec<AnalysisType>,
-
-    /// Maximum trigger rate per second
-    pub max_trigger_rate: u32,
-
-    /// Maximum concurrent analysis tasks
-    pub max_concurrent_analysis: u32,
-
-    /// Supported priority levels
-    pub supported_priorities: Vec<TriggerPriority>,
-
-    /// Resource limits for analysis
-    pub resource_limits: TriggerResourceLimits,
-}
-
-/// Resource limits for trigger analysis.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerResourceLimits {
-    /// Maximum memory usage per analysis task (bytes)
-    pub max_memory_per_task: u64,
-
-    /// Maximum analysis time per task (milliseconds)
-    pub max_analysis_time_ms: u64,
-
-    /// Maximum queue depth for pending analysis
-    pub max_queue_depth: usize,
-}
-
-impl Default for TriggerResourceLimits {
-    fn default() -> Self {
-        Self {
-            max_memory_per_task: 100 * 1024 * 1024, // 100MB
-            max_analysis_time_ms: 30_000,           // 30 seconds
-            max_queue_depth: 1000,
-        }
-    }
-}
-
-/// Priority-based trigger queue for managing analysis requests.
-#[derive(Debug)]
-pub struct PriorityTriggerQueue {
-    /// High priority queue (Critical, High)
-    high_priority: VecDeque<TriggerRequest>,
-
-    /// Normal priority queue (Normal, Low)
-    normal_priority: VecDeque<TriggerRequest>,
-
-    /// Maximum queue size per priority level
-    max_queue_size: usize,
-
-    /// Backpressure threshold (percentage of max size)
-    backpressure_threshold: f32,
-
-    /// Queue statistics
-    stats: QueueStatistics,
-}
-
-/// Queue statistics for monitoring.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct QueueStatistics {
-    /// Total triggers enqueued
-    pub total_enqueued: u64,
-
-    /// Total triggers processed
-    pub total_processed: u64,
-
-    /// Triggers dropped due to backpressure
-    pub dropped_backpressure: u64,
-
-    /// Triggers dropped due to queue full
-    pub dropped_queue_full: u64,
-
-    /// Current queue depths
-    pub high_priority_depth: usize,
-    pub normal_priority_depth: usize,
-}
-
-impl PriorityTriggerQueue {
-    /// Creates a new priority trigger queue.
-    ///
-    /// # Examples
-    ///
-    /// ```
-    /// use collector_core::trigger::PriorityTriggerQueue;
-    ///
-    /// let queue = PriorityTriggerQueue::new(1000, 0.8);
-    /// ```
-    pub fn new(max_queue_size: usize, backpressure_threshold: f32) -> Self {
-        Self {
-            high_priority: VecDeque::with_capacity(max_queue_size / 2),
-            normal_priority: VecDeque::with_capacity(max_queue_size / 2),
-            max_queue_size,
-            backpressure_threshold,
-            stats: QueueStatistics::default(),
-        }
-    }
-
-    /// Enqueues a trigger request with priority-based routing.
-    pub fn enqueue(&mut self, trigger: TriggerRequest) -> Result<(), TriggerError> {
-        self.stats.total_enqueued += 1;
-
-        // Check for backpressure
-        if self.is_backpressure_active() {
-            // Drop low priority triggers during backpressure
-            if matches!(trigger.priority, TriggerPriority::Low) {
-                self.stats.dropped_backpressure += 1;
-                warn!(
-                    trigger_id = %trigger.trigger_id,
-                    priority = ?trigger.priority,
-                    "Dropping low priority trigger due to backpressure"
-                );
-                return Err(TriggerError::BackpressureActive);
-            }
-        }
-
-        // Route to appropriate queue based on priority
-        match trigger.priority {
-            TriggerPriority::Critical | TriggerPriority::High => {
-                if self.high_priority.len() >= self.max_queue_size / 2 {
-                    self.stats.dropped_queue_full += 1;
-                    return Err(TriggerError::QueueFull("high_priority".to_owned()));
-                }
-                self.high_priority.push_back(trigger);
-                self.stats.high_priority_depth = self.high_priority.len();
-            }
-            TriggerPriority::Normal | TriggerPriority::Low => {
-                if self.normal_priority.len() >= self.max_queue_size / 2 {
-                    self.stats.dropped_queue_full += 1;
-                    return Err(TriggerError::QueueFull("normal_priority".to_owned()));
-                }
-                self.normal_priority.push_back(trigger);
-                self.stats.normal_priority_depth = self.normal_priority.len();
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Dequeues the next trigger request, prioritizing high priority items.
-    pub fn dequeue(&mut self) -> Option<TriggerRequest> {
-        // Always process high priority first
-        if let Some(trigger) = self.high_priority.pop_front() {
-            self.stats.total_processed += 1;
-            self.stats.high_priority_depth = self.high_priority.len();
-            return Some(trigger);
-        }
-
-        // Then process normal priority
-        if let Some(trigger) = self.normal_priority.pop_front() {
-            self.stats.total_processed += 1;
-            self.stats.normal_priority_depth = self.normal_priority.len();
-            return Some(trigger);
-        }
-
-        None
-    }
-
-    /// Checks if backpressure is currently active.
-    pub fn is_backpressure_active(&self) -> bool {
-        let total_depth = self.high_priority.len() + self.normal_priority.len();
-        let threshold = (self.max_queue_size as f32 * self.backpressure_threshold) as usize;
-        total_depth >= threshold
-    }
-
-    /// Returns current queue statistics.
-    pub fn get_statistics(&self) -> QueueStatistics {
-        let mut stats = self.stats.clone();
-        stats.high_priority_depth = self.high_priority.len();
-        stats.normal_priority_depth = self.normal_priority.len();
-        stats
-    }
-
-    /// Returns the total number of pending triggers.
-    pub fn len(&self) -> usize {
-        self.high_priority.len() + self.normal_priority.len()
-    }
-
-    /// Returns true if the queue is empty.
-    pub fn is_empty(&self) -> bool {
-        self.high_priority.is_empty() && self.normal_priority.is_empty()
-    }
-}
-
-/// SQL-to-IPC integration for trigger condition evaluation.
-#[derive(Debug)]
-pub struct SqlTriggerEvaluator {
-    /// SQL dialect for parsing
-    dialect: GenericDialect,
-
-    /// Compiled trigger conditions mapped by collector
-    compiled_conditions: HashMap<String, Vec<CompiledTriggerCondition>>,
-
-    /// Performance statistics
-    evaluation_stats: SqlEvaluationStats,
-}
-
-/// Compiled trigger condition for efficient evaluation.
-#[derive(Debug, Clone)]
-pub struct CompiledTriggerCondition {
-    /// Original condition
-    pub condition: TriggerCondition,
-
-    /// Compiled SQL predicate (if applicable)
-    pub compiled_predicate: Option<sqlparser::ast::Expr>,
-
-    /// Evaluation statistics
-    pub stats: ConditionEvaluationStats,
-}
-
-/// Statistics for SQL trigger evaluation.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct SqlEvaluationStats {
-    /// Total evaluations performed
-    pub total_evaluations: u64,
-
-    /// Successful matches
-    pub successful_matches: u64,
-
-    /// Evaluation errors
-    pub evaluation_errors: u64,
-
-    /// Average evaluation time (microseconds)
-    pub avg_evaluation_time_us: u64,
-
-    /// SQL parsing errors
-    pub sql_parsing_errors: u64,
-}
-
-/// Statistics for individual condition evaluation.
-#[derive(Debug, Clone, Default)]
-pub struct ConditionEvaluationStats {
-    /// Number of times this condition was evaluated
-    pub evaluation_count: u64,
-
-    /// Number of times this condition matched
-    pub match_count: u64,
-
-    /// Number of times this condition did not match
-    pub no_match_count: u64,
-
-    /// Total evaluation time (microseconds)
-    pub total_evaluation_time_us: u64,
-
-    /// Last evaluation timestamp
-    pub last_evaluation: Option<SystemTime>,
-
-    /// Last time this condition matched
-    pub last_match: Option<SystemTime>,
-
-    /// Last time this condition did not match
-    pub last_no_match: Option<SystemTime>,
-}
-
-impl SqlTriggerEvaluator {
-    /// Creates a new SQL trigger evaluator.
-    pub fn new() -> Self {
-        Self {
-            dialect: GenericDialect {},
-            compiled_conditions: HashMap::new(),
-            evaluation_stats: SqlEvaluationStats::default(),
-        }
-    }
-
-    /// Registers trigger conditions for a collector.
-    pub fn register_collector_conditions(
-        &mut self,
-        collector_id: String,
-        conditions: Vec<TriggerCondition>,
-    ) -> Result<(), TriggerError> {
-        let mut compiled_conditions = Vec::new();
-
-        for condition in conditions {
-            let compiled = self.compile_condition(condition.clone())?;
-            compiled_conditions.push(compiled);
-        }
-
-        self.compiled_conditions
-            .insert(collector_id, compiled_conditions);
-        Ok(())
-    }
-
-    /// Compiles a trigger condition for efficient evaluation.
-    fn compile_condition(
-        &self,
-        condition: TriggerCondition,
-    ) -> Result<CompiledTriggerCondition, TriggerError> {
-        #[allow(clippy::wildcard_enum_match_arm)]
-        let compiled_predicate = match &condition.condition_type {
-            ConditionType::Custom(sql_expr) => {
-                // Parse SQL expression for custom conditions
-                match self.parse_sql_expression(sql_expr) {
-                    Ok(expr) => Some(expr),
-                    Err(e) => {
-                        warn!(
-                            condition_id = %condition.id,
-                            error = %e,
-                            "Failed to compile SQL expression for trigger condition"
-                        );
-                        None
-                    }
-                }
-            }
-            _ => None, // Built-in conditions don't need SQL compilation
-        };
-
-        Ok(CompiledTriggerCondition {
-            condition,
-            compiled_predicate,
-            stats: ConditionEvaluationStats::default(),
-        })
-    }
-
-    /// Parses a SQL WHERE clause expression from a string into an AST.
-    ///
-    /// This function is designed to parse simple SQL predicate expressions suitable for
-    /// filtering process data. It has specific limitations:
-    ///
-    /// - **Predicate-only**: The input string must represent only the WHERE clause content,
-    ///   not a full `SELECT` statement.
-    /// - **No table references**: Expressions should not contain table-qualified identifiers
-    ///   (e.g., `processes.name`). Only unqualified column names (e.g., `name`) are supported.
-    /// - **No subqueries**: Subqueries (e.g., `(SELECT ...)`) are not allowed.
-    /// - **Unqualified column names**: All column names are assumed to refer to fields
-    ///   within the `ProcessInfo` structure.
-    ///
-    /// # Arguments
-    /// * `sql_expr` - The SQL WHERE clause expression string.
-    ///
-    /// # Returns
-    /// A `Result` containing the parsed `Expr` AST node on success, or a `TriggerError` on failure.
-    ///
-    /// # Errors
-    /// Returns `TriggerError::SqlParsingError` if the expression is invalid or contains
-    /// disallowed patterns.
-    ///
-    /// # Examples
-    ///
-    /// **Valid expressions:**
-    /// ```ignore
-    /// let expr = parse_sql_expression("pid > 1000 AND name LIKE 'chrome%'")?;
-    /// ```
-    ///
-    /// **Invalid expressions (will return an error):**
-    /// ```ignore
-    /// // Contains a disallowed "SELECT" keyword
-    /// parse_sql_expression("pid IN (SELECT id FROM other_table)"); // Error
-    ///
-    /// // Contains a table-qualified identifier
-    /// parse_sql_expression("processes.pid > 100"); // Error
-    /// ```
-    fn parse_sql_expression(&self, sql_expr: &str) -> Result<sqlparser::ast::Expr, TriggerError> {
-        // Lightweight validation to reject disallowed patterns before parsing
-        let expr_upper = sql_expr.to_uppercase();
-        if expr_upper.contains("SELECT")
-            || expr_upper.contains("FROM")
-            || sql_expr.contains('.')
-            || (sql_expr.contains('(') && expr_upper.contains("SELECT"))
-        {
-            return Err(TriggerError::SqlParsingError(format!(
-                "Expression contains disallowed patterns (e.g., SELECT, FROM, table-qualified identifiers, subqueries): {sql_expr}"
-            )));
-        }
-
-        // Wrap the expression in a SELECT statement for parsing
-        let sql = format!("SELECT * FROM dummy WHERE {sql_expr}");
-
-        let statements = Parser::parse_sql(&self.dialect, &sql).map_err(|e| {
-            TriggerError::SqlParsingError(format!("Failed to parse SQL expression: {e}"))
-        })?;
-
-        if let Some(sqlparser::ast::Statement::Query(query)) = statements.first()
-            && let sqlparser::ast::SetExpr::Select(select) = query.body.as_ref()
-            && let Some(selection) = &select.selection
-        {
-            return Ok(selection.clone());
-        }
-
-        Err(TriggerError::SqlParsingError(
-            "Could not extract WHERE clause expression from SQL statement.".to_owned(),
-        ))
-    }
-
-    /// Evaluates trigger conditions against process data using SQL-like logic.
-    pub fn evaluate_sql_triggers(
-        &mut self,
-        collector_id: &str,
-        process_data: &ProcessTriggerData,
-    ) -> Result<Vec<TriggerRequest>, TriggerError> {
-        let start_time = SystemTime::now();
-        self.evaluation_stats.total_evaluations += 1;
-
-        // Get conditions reference to avoid expensive clone
-        let Some(conditions) = self.compiled_conditions.get(collector_id) else {
-            return Ok(Vec::new()); // No conditions registered for this collector
-        };
-
-        let mut triggers = Vec::new();
-        let mut match_results = Vec::new();
-
-        for (index, compiled_condition) in conditions.iter().enumerate() {
-            let condition_start = SystemTime::now();
-
-            let matches = match Self::evaluate_compiled_condition(compiled_condition, process_data)
-            {
-                Ok(matches) => matches,
-                Err(e) => {
-                    self.evaluation_stats.evaluation_errors += 1;
-                    warn!(
-                        condition_id = %compiled_condition.condition.id,
-                        collector_id = %collector_id,
-                        error = %e,
-                        "Failed to evaluate trigger condition"
-                    );
-                    match_results.push((index, false, condition_start));
-                    continue;
-                }
-            };
-
-            match_results.push((index, matches, condition_start));
-
-            if matches {
-                self.evaluation_stats.successful_matches += 1;
-
-                // Create trigger request
-                let trigger =
-                    Self::create_sql_trigger_request(&compiled_condition.condition, process_data)?;
-                triggers.push(trigger);
-            }
-        }
-
-        // Update condition statistics after the loop to avoid borrow conflicts
-        for (index, matches, _condition_start) in match_results {
-            if let Some(conditions_mut) = self.compiled_conditions.get_mut(collector_id)
-                && let Some(compiled_condition_mut) = conditions_mut.get_mut(index)
-            {
-                compiled_condition_mut.stats.evaluation_count += 1;
-                compiled_condition_mut.stats.last_evaluation = Some(SystemTime::now());
-
-                if matches {
-                    compiled_condition_mut.stats.match_count += 1;
-                    compiled_condition_mut.stats.last_match = Some(SystemTime::now());
-                } else {
-                    compiled_condition_mut.stats.no_match_count += 1;
-                    compiled_condition_mut.stats.last_no_match = Some(SystemTime::now());
-                }
-            }
-        }
-
-        // Update overall evaluation time
-        if let Ok(elapsed) = start_time.elapsed() {
-            let elapsed_us = elapsed.as_micros() as u64;
-            let total_evals = self.evaluation_stats.total_evaluations;
-            self.evaluation_stats.avg_evaluation_time_us =
-                (self.evaluation_stats.avg_evaluation_time_us * (total_evals - 1) + elapsed_us)
-                    / total_evals;
-        }
-
-        Ok(triggers)
-    }
-
-    /// Evaluates a compiled trigger condition against process data.
-    fn evaluate_compiled_condition(
-        compiled_condition: &CompiledTriggerCondition,
-        data: &ProcessTriggerData,
-    ) -> Result<bool, TriggerError> {
-        match &compiled_condition.condition.condition_type {
-            ConditionType::ProcessNamePattern(pattern) => Ok(data.name.contains(pattern)),
-            ConditionType::ExecutablePathPattern(pattern) => Ok(data
-                .executable_path
-                .as_ref()
-                .is_some_and(|path| path.contains(pattern))),
-            ConditionType::MissingExecutable => Ok(!data.file_exists),
-            ConditionType::HashMismatch => {
-                // This would integrate with actual hash verification logic
-                Ok(false)
-            }
-            ConditionType::SuspiciousParentChild => {
-                // This would integrate with parent-child relationship analysis
-                Ok(false)
-            }
-            ConditionType::ResourceAnomaly {
-                cpu_threshold,
-                memory_threshold,
-            } => {
-                let cpu_anomaly = data.cpu_usage.is_some_and(|cpu| cpu > *cpu_threshold);
-                let memory_anomaly = data.memory_usage.is_some_and(|mem| mem > *memory_threshold);
-                Ok(cpu_anomaly || memory_anomaly)
-            }
-            ConditionType::Custom(_sql_expr) => {
-                // TODO: Implement SQL evaluation for custom conditions.
-                // For now, custom conditions are not supported and always evaluate to false.
-                // See issue #000 for tracking.
-                warn!("Custom SQL conditions are not yet supported and always evaluate to false.");
-                Ok(false)
-            }
-        }
-    }
-
-    /// Creates a trigger request from a matched condition and process data.
-    fn create_sql_trigger_request(
-        condition: &TriggerCondition,
-        data: &ProcessTriggerData,
-    ) -> Result<TriggerRequest, TriggerError> {
-        let trigger_id = Uuid::new_v4().to_string();
-        let correlation_id = Uuid::new_v4().to_string();
-        let timestamp = SystemTime::now();
-
-        let mut metadata = HashMap::new();
-        metadata.insert("condition_id".to_owned(), condition.id.clone());
-        metadata.insert("source_pid".to_owned(), data.pid.to_string());
-        metadata.insert("evaluation_method".to_owned(), "sql_trigger".to_owned());
-
-        if let Some(ref path) = data.executable_path {
-            metadata.insert("executable_path".to_owned(), path.clone());
-        }
-
-        Ok(TriggerRequest {
-            trigger_id,
-            target_collector: condition.target_collector.clone(),
-            analysis_type: condition.analysis_type.clone(),
-            priority: condition.priority.clone(),
-            target_pid: Some(data.pid),
-            target_path: data.executable_path.clone(),
-            correlation_id,
-            metadata,
-            timestamp,
-        })
-    }
-
-    /// Returns SQL evaluation statistics.
-    pub fn get_evaluation_stats(&self) -> SqlEvaluationStats {
-        self.evaluation_stats.clone()
-    }
-
-    /// Returns condition-specific statistics for a collector.
-    pub fn get_condition_stats(
-        &self,
-        collector_id: &str,
-    ) -> Vec<(String, ConditionEvaluationStats)> {
-        self.compiled_conditions
-            .get(collector_id)
-            .map(|conditions| {
-                conditions
-                    .iter()
-                    .map(|c| (c.condition.id.clone(), c.stats.clone()))
-                    .collect()
-            })
-            .unwrap_or_default()
-    }
-}
-
-impl Default for SqlTriggerEvaluator {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Trigger system statistics for monitoring.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct TriggerStatistics {
-    /// Number of registered trigger conditions
-    pub registered_conditions: usize,
-
-    /// Number of pending triggers
-    pub pending_triggers: usize,
-
-    /// Size of deduplication cache
-    pub deduplication_cache_size: usize,
-
-    /// Number of rate limit states tracked
-    pub rate_limit_states: usize,
-
-    /// Number of registered collector capabilities
-    pub registered_capabilities: usize,
-
-    /// Priority queue statistics
-    pub queue_stats: QueueStatistics,
-
-    /// SQL evaluation statistics
-    pub sql_evaluation_stats: SqlEvaluationStats,
-
-    /// Trigger emission statistics
-    pub emission_stats: TriggerEmissionStats,
-}
-
-/// Statistics for trigger request emission and routing.
-#[derive(Debug, Clone, Default, Serialize, Deserialize)]
-pub struct TriggerEmissionStats {
-    /// Total trigger requests emitted
-    pub total_emitted: u64,
-
-    /// Successful emissions to event bus
-    pub successful_emissions: u64,
-
-    /// Failed emissions due to validation errors
-    pub validation_failures: u64,
-
-    /// Failed emissions due to event bus errors
-    pub event_bus_failures: u64,
-
-    /// Trigger requests that timed out
-    pub timeouts: u64,
-
-    /// Average emission latency in milliseconds
-    pub avg_emission_latency_ms: f64,
-
-    /// Trigger requests currently pending response
-    pub pending_responses: usize,
-}
-
 /// Timeout tracking for trigger requests.
 #[derive(Debug, Clone)]
 struct TriggerTimeout {
@@ -1948,56 +1165,6 @@ struct TriggerTimeout {
     /// Correlation ID for forensic tracking
     #[allow(dead_code)]
     correlation_id: String,
-}
-
-/// Errors that can occur in the trigger system.
-#[derive(Debug, thiserror::Error)]
-#[non_exhaustive]
-pub enum TriggerError {
-    #[error("Lock error on {0}")]
-    LockError(String),
-
-    #[error("Configuration error: {0}")]
-    ConfigError(String),
-
-    #[error("Condition evaluation error: {0}")]
-    EvaluationError(String),
-
-    #[error("Trigger generation error: {0}")]
-    GenerationError(String),
-
-    #[error("SQL parsing error: {0}")]
-    SqlParsingError(String),
-
-    #[error("Capability validation error: {0}")]
-    CapabilityValidationError(String),
-
-    #[error("Queue full: {0}")]
-    QueueFull(String),
-
-    #[error("Backpressure active - dropping low priority triggers")]
-    BackpressureActive,
-
-    #[error("Collector not found: {0}")]
-    CollectorNotFound(String),
-
-    #[error("Invalid trigger priority: {0:?}")]
-    InvalidPriority(TriggerPriority),
-
-    #[error("Trigger emission failed: {0}")]
-    EmissionError(String),
-
-    #[error("Event bus error: {0}")]
-    EventBusError(String),
-
-    #[error("Trigger validation failed: {0}")]
-    ValidationError(String),
-
-    #[error("Trigger timeout: {0}")]
-    TimeoutError(String),
-
-    #[error("Correlation tracking error: {0}")]
-    CorrelationError(String),
 }
 
 #[cfg(test)]
@@ -2039,6 +1206,12 @@ pub enum TriggerError {
 )]
 mod tests {
     use super::*;
+    use crate::event::{AnalysisType, TriggerPriority};
+    use crate::trigger::queue::PriorityTriggerQueue;
+    use crate::trigger::types::{
+        ConditionType, DeduplicationKey, ProcessTriggerData, TriggerCapabilities, TriggerCondition,
+        TriggerResourceLimits,
+    };
 
     fn create_test_process_data() -> ProcessTriggerData {
         ProcessTriggerData {
